feat: nix flake
