+++ .gitignore
@@ -23,3 +23,4 @@ bin/
 
 # OS files
 .DS_Store
+result
